+++ ICD.Profound.ConnectPRO/Themes/UserInterface/IPresenters/WebConference/ActiveMeeting/IWtcParticipantControlsPresenter.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using ICD.Connect.Conferencing.Participants;
+﻿using ICD.Connect.Conferencing.Participants;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.WebConference.ActiveMeeting;
 
 namespace ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.WebConference.ActiveMeeting