+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/Common/ReferencedSchedulePresenter.cs
@@ -7,7 +7,6 @@ using ICD.Common.Utils.Extensions;
 using ICD.Connect.Calendaring;
 using ICD.Connect.Calendaring.Booking;
 using ICD.Connect.Conferencing.Controls.Dialing;
-using ICD.Connect.Conferencing.Utils;
 using ICD.Connect.Partitioning.Rooms;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.Common;