+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/WebConference/ActiveMeeting/WtcActiveMeetingPresenter.cs
@@ -9,9 +9,7 @@ using ICD.Connect.Conferencing.EventArguments;
 using ICD.Connect.Conferencing.Participants;
 using ICD.Connect.Conferencing.Zoom.Components.Call;
 using ICD.Connect.UI.Mvp.Presenters;
-using ICD.Connect.UI.Utils;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters;
-using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.WebConference;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.WebConference.ActiveMeeting;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.WebConference.Contacts;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews;
@@ -19,7 +17,7 @@ using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.WebConference.ActiveMe
 
 namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.WebConference.ActiveMeeting
 {
-	public class WtcActiveMeetingPresenter : AbstractWtcPresenter<IWtcActiveMeetingView>, IWtcActiveMeetingPresenter
+	public sealed class WtcActiveMeetingPresenter : AbstractWtcPresenter<IWtcActiveMeetingView>, IWtcActiveMeetingPresenter
 	{
 		private readonly SafeCriticalSection m_RefreshSection;
 		private readonly WtcReferencedParticipantPresenterFactory m_PresenterFactory;
@@ -301,7 +299,7 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.WebConference.
 
 		protected override void ViewOnPreVisibilityChanged(object sender, BoolEventArgs args)
 		{
-			base.ViewOnVisibilityChanged(sender, args);
+			base.ViewOnPreVisibilityChanged(sender, args);
 
 			SelectedParticipant = null;
 			if (!args.Data)