+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Views/WebConference/WtcStartMeetingView.cs
@@ -2,7 +2,6 @@
 using ICD.Common.Utils.EventArguments;
 using ICD.Common.Utils.Extensions;
 using ICD.Connect.Panels;
-using ICD.Connect.UI.Controls;
 using ICD.Connect.UI.EventArguments;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.WebConference;